Bug in here
If you change the volume of an existing program(A) and then select another program(B), the volume value of the existing program(A) is applied to the other program(B).
Before using the value applied to the previous program(A), the volume of the changed program(B) must be read first.